Moved content/ to root
