+++ components/visgl/pages/showcases.tsx
@@ -1,9 +1,9 @@
-"use client"
+'use client';
 
 import React from 'react';
 import Showcase from '../components/showcase';
 
-import showcasesYaml from '../content/showcases.json';
+import showcasesYaml from '@/content/showcases.json';
 export type ShowcaseEntry = {
   description: string;
   image: string;
@@ -11,7 +11,7 @@ export type ShowcaseEntry = {
   name: string;
   url: string;
 };
-declare module '../content/showcases.json' {
+declare module '@/content/showcases.json' {
   const showcases: ShowcaseEntry[];
 }
