+++ content/deprecated/blog.yaml
